backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9
pymupdf==1.24.10
Pillow==10.4.0
rapidocr-onnxruntime==1.3.25
onnxruntime==1.18.0